requirements.txt
anyio==3.5.0
asgiref==3.5.1
attrs==21.4.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.1
fastapi==0.76.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
jwcrypto==1.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.18.0
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
